Use Floats instead of Doubles for node parameters

+++ src/main/kotlin/laser/Point.kt
@@ -4,22 +4,22 @@ import kotlin.math.cos
 import kotlin.math.sin
 
 data class Point(
-    var x: Double,
-    var y: Double,
+    var x: Float,
+    var y: Float,
 ) {
-    fun offset(x: Double, y: Double) {
+    fun offset(x: Float, y: Float) {
         this.x += x
         this.y += y
     }
 
-    fun rotate(angle: Double, anchorX: Double, anchorY: Double) {
+    fun rotate(angle: Float, anchorX: Float, anchorY: Float) {
         if (angle % 1.0 == 0.0) return
         val radians = angle * 2 * Math.PI
         val translatedX = this.x - anchorX
         val translatedY = this.y - anchorY
         val rotatedX = translatedX * cos(radians) - translatedY * sin(radians)
         val rotatedY = translatedX * sin(radians) + translatedY * cos(radians)
-        this.x = rotatedX + anchorX
-        this.y = rotatedY + anchorY
+        this.x = (rotatedX + anchorX).toFloat()
+        this.y = (rotatedY + anchorY).toFloat()
     }
 }
